Initial scrypt support
This only works with an as of yet unreleased version of scrypt.
Furthermore there seems to be some issue with Python 3.x so just as with
bcrypt there is no scrypt support for Python 3.x either.

Furthermore this just uses parameters that are probably acceptable for
web applications right now but there is no mechanism to determine useful
parameters right now and I'm unsure on how to go about that.